+++ CHANGELOG.md
@@ -4,6 +4,49 @@ _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob
 
 - **[UPDATE]** Updated to Kotlin 1.5.31.
 - **[UPDATE]** Updated to Kotlin Coroutines 1.5.2.
+- **[FEATURE]** (`ktx-reflect`) Added a new module with reflection utilities.
+  - `Reflection` annotation, which requires opt-in, allows marking functionalities that rely on reflection.
+  - `ReflectedClass` is an inlined wrapper for `Class` that allows to easily use LibGDX reflection API.
+    - `ClassReflection` methods exposed as `ReflectedClass` read-only properties:
+      - `simpleName`
+      - `isMemberClass`
+      - `isStatic`
+      - `isArray`
+      - `isPrimitive`
+      - `isEnum`
+      - `isAnnotation`
+      - `isInterface`
+      - `isAbstract`
+      - `componentType`
+      - `constructors`
+      - `methods`
+      - `declaredMethods`
+      - `fields`
+      - `declaredFields`
+      - `annotations`
+      - `declaredAnnotations`
+      - `interfaces`
+      - `enumConstants`
+    - `ClassReflection` methods exposed as `ReflectedClass` methods with improved Kotlin class handling:
+      - `isInstance`
+      - `isAssignableFrom`
+      - `newInstance`
+      - `getConstructor`
+      - `getDeclaredConstructor`
+      - `getMethod`
+      - `getDeclaredMethod`
+      - `getField`
+      - `getDeclaredField`
+      - `isAnnotationPresent`
+      - `getAnnotation`
+      - `getDeclaredAnnotation`
+    - Additional `ReflectedClass` utilities:
+      - `constructor` property allows to extract the only class constructor if it is available.
+      - `newArrayInstance` method allows to create a type array of the given size using `ArrayReflection`.
+      - `KClass` constructor allows creating `ReflectedClass` from a `KClass`.
+  - `reflect<Type>()` method allows to wrap a selected class with `ReflectedClass`.
+  - `reflect(String)` method allows to find a selected class by qualified name and wrap it with `ReflectedClass`.
+  - `Annotation.get` extension method with a reified type allowing to get an instance of the annotation.
 
 #### 1.10.0-b2